Fix intensity scale selection bug
It appears that the code for intensity scale selection had a bug where
it used the updated scale minimum when computing the update to the scale
maximum. The result being we could end up with a scale maximum larger
than it should be if an adjustment was made to the scale minimum. We fix
this by using temporaries to store the results of the update minimum and
maximum before actually storing the updated scale values. As we don't
use the temporaries in our computation, we won't re-encounter the same
bug.