Bust GitHub's badge cache
Appears that GitHub's cache is stale and nothing else is working. This
attempts to fix the issue by changing the URL. Here we pass unused
arguments to the outdated badges. While the arguments do nothing for the
services, this does change the URL, which forces camo, the caching
service, to retrieve the badges as these URLs are not already cached.
Also it should allow camo to drop the existing badges from cache during
garbage collection. That way we can return to the original URLs and get
new badges into cache.

[ci skip]