Fix a bare `except` clause by requiring `Exception`
There is no reason we need to catch an arbitrary exception here. We
should be getting a NumPy array, which we will need to index. The index
could be out of bounds, which would be an `IndexError`. There is a
remote possibility that the coordinates provides are not `float` for
some reason. So coercing them to `int`s may fail, which would be a
`ValueError`. To keep this general, just make the bare `except` catch
any `Exception`, which should handle both of these cases.