Cache the current frame for fast access
When working with large data on a slow filesystem, re-reading the
current displayed image is unnecessarily slow. This happens when
checking the intensity of different coordinates for instance (via
mouseover events). To improve performance, we can store the currently
displayed frame and simply access it if the user has not selected a
different frame.

We accomplish this here by storing the frame number (time) and the
current frame's data. These are both initialized to be `None` so that
they will fail cache hits the first time.  Within `get_image`, checks
are added to see if the current frame number (time) and requested one
are the same. If not, the new frame is loaded and the cache values
updated once loading completes. This ensures that some raised exception
doesn't create an incorrectly determined cached state. Whether or not
the frame is current or new, the frame is returned.